+++ milvus-20231012/ChangeLog.md
@@ -1,3 +1,10 @@
+2025-03-19 Version: 1.1.0
+- Support API CreateDefaultRole.
+- Update API GetInstanceDetail: update response param.
+- Update API ListInstances: add param Tag.
+- Update API ListInstances: update response param.
+
+
 2024-12-24 Version: 1.0.7
 - Update API GetInstanceDetail: update response param.
 - Update API ListInstances: update response param.